Track path conditions
Including handling `if *`; turns out this basically requires the same
infrastructure as other uses of havoc, we just evaluate * as usual when
encountering the `if` and create a nullary relation for which branch
we're in. This is needed since there can be assumes/assignments in each
branch and these have to all be kept consistent with which path we're
in.